+++ client/components/pages/ViewDocumentPage.jsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router';
 import toastr from 'toastr';
 import Divider from 'material-ui/Divider';
 import handleError from '../../utils/errorHandler';
-import Sidebar from '../layouts/Sidebar.jsx';
+import Sidebar from '../layouts/Sidebar';
 import { getDocument, deleteDocument } from '../../actions/documentActions';
 
 /**
@@ -22,7 +22,7 @@ export class ViewDocumentPage extends React.Component {
    * Calls actions to get the document by id
    * @memberOf ViewDocumentPage
    */
-  componentWillMount() {
+  componentDidMount() {
     this.props.getDocument(this.props.params.id);
   }
 
@@ -31,22 +31,22 @@ export class ViewDocumentPage extends React.Component {
    * @memberOf ViewDocumentPage
    */
   deleteDocument = () => {
-    // swal({
-    //   title: 'Are you sure?',
-    //   text: 'You will not be able to recover this document!',
-    //   type: 'warning',
-    //   showCancelButton: true,
-    //   confirmButtonColor: '#DD6B55',
-    //   confirmButtonText: 'Yes, delete it!',
-    //   cancelButtonText: 'No, cancel this!'
-    // })
-    // .then((isConfirm) => {
-    //   if (isConfirm) {
-    this.delete();
-    //   }
-    // })
-    // .catch((er) =>
-    //   swal('Cancelled', 'The document is safe :)', 'error'));
+    swal({
+      title: 'Are you sure?',
+      text: 'You will not be able to recover this document!',
+      type: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#DD6B55',
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, cancel this!'
+    })
+    .then((isConfirm) => {
+      if (isConfirm) {
+        this.delete();
+      }
+    })
+    .catch((er) =>
+      swal('Cancelled', 'The document is safe :)', 'error'));
   }
 
   /**
@@ -63,7 +63,7 @@ export class ViewDocumentPage extends React.Component {
    * @memberOf ViewDocumentPage
    */
   redirect() {
-    // swal('Deleted!', 'This document has been deleted.', 'success');
+    swal('Deleted!', 'This document has been deleted.', 'success');
     this.context.router.push('/home');
   }
